Added basic task system
The file system has been rounded out a bit more and now has the ability
to handle relative, absolute, and network based files. Uris are probably
next there.

Also added a very basic "task" system. Basically just loads small tasks
and calls Run at the appropriate times. Much of the platform specific
configuration should be able to tie into it.

Next biit should be interesting as it will be more of the ASP.Net/MVC
specific items (minification, etc).

+++ Batman.Core.FileSystem.Local/Directory.cs
@@ -128,6 +128,8 @@ namespace Batman.Core.FileSystem.Local
 
         public void Delete()
         {
+            if (!Exists)
+                return;
             InternalDirectory.DeleteAll();
             InternalDirectory.Refresh();
         }
